requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
asyncmy==0.2.10
click==8.1.8
comm==0.2.2
databases==0.9.0
debugpy==1.8.12
decorator==5.1.1
executing==2.2.0
fastapi==0.115.8
greenlet==3.1.1
grpcio==1.70.0
grpcio-tools==1.70.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httptools==0.6.4
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
mariadb==1.1.11
matplotlib-inline==0.1.7
mysqlclient==2.2.7
mysqlclient-pool==0.1.2
nest-asyncio==1.6.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
priority==2.0.0
prompt_toolkit==3.0.50
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==26.2.1
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
stack-data==0.6.3
starlette==0.45.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
uvicorn==0.34.0
uvloop==0.21.0
wcwidth==0.2.13
wsproto==1.2.0
